moved IMessage to a separate file
