+++ src/components/Player/Player.tsx
@@ -15,7 +15,7 @@ import {
 import playerIdToString from "../../lib/playerIdToString";
 import playerStringToId from "../../lib/playerStringToId";
 import { IPlayer, IState } from "../../store/types";
-import { IMessage } from "../Game/onMessage";
+import { IMessage } from "../Game/types/IMessage";
 import { Possibilities, PlayerActions, GameTurns } from "../../lib/constants";
 import notifications from "../../config/notifications.json";
 import sounds from "../../sounds/sounds";
